DocumentChatbot/requirements.txt
# Flask
flask

# Document processing
PyPDF2
python-docx
openpyxl
pandas

# HTTP requests for Ollama API
requests
